+++ src/components/Button/index.js
@@ -58,6 +58,7 @@ Button.propTypes = {
     'outline-reverse',
     'plain',
     'toggle',
+    'underline',
     'unstyled',
   ]),
   /** Hook for automated JavaScript tests */